+++ backend/src/entity/User.ts
@@ -28,3 +28,4 @@ export class User extends BaseEntity{
 
 
 }
+
