+++ backend/tsconfig.json
@@ -11,5 +11,5 @@
       "emitDecoratorMetadata": true,
       "experimentalDecorators": true,
       "sourceMap": true
-   }
+    },
 }
